Update and rename createuser-elvir.cs to create-user&delete-user.cs

+++ rsa/create-user&delete-user.cs
@@ -3,6 +3,26 @@ using System.Security.Cryptography;
 using System.IO;
 using System.Text;
 
+
+/*
+ * C:\Users\HP\source\repos\rsaa\rsaa\bin\Debug\netcoreapp3.1>rsaa create-user Fiek.1
+Argumenti i dyte duhet te jete A-Z,a-z,0-9.
+
+C:\Users\HP\source\repos\rsaa\rsaa\bin\Debug\netcoreapp3.1>rsaa create-user Fiek1
+Eshte krijua celsi privat 'keys/Fiek1.xml'
+Eshte krijua celse publik 'keys/Fiek1.pub.xml'
+
+C:\Users\HP\source\repos\rsaa\rsaa\bin\Debug\netcoreapp3.1>rsaa create-user Fiek1
+Gabim: Celesi 'Fiek1' ekziston paraprakisht.
+
+C:\Users\HP\source\repos\rsaa\rsaa\bin\Debug\netcoreapp3.1>rsaa delete-user Fiek1
+Eshte larguar celsi privat 'keys/Fiek1.xml'
+Eshte larguar celse publik 'keys/Fiek1.pub.xml'
+
+C:\Users\HP\source\repos\rsaa\rsaa\bin\Debug\netcoreapp3.1>rsaa delete-user Fiek1
+Gabim: Celesi 'Fiek1' ekziston paraprakisht.
+
+*/
 namespace RsaSignature
 {
     class Program
@@ -30,46 +50,69 @@ namespace RsaSignature
                 StringBuilder sb = new StringBuilder();
                 for(int i = 0; i < command2.Length; i++)
                 {
-                    if (command2[i] >= 'a' && command2[i] <= 'z' && command2[i] >= 'A' && command2[i] >= 'Z' && command2[i] >= '0' && command2[i] >= '9')
+                    if ((command2[i] >= '0' && command2[i] <= '9') || (command2[i] >= 'A' && command2[i] <= 'Z' )|| (command2[i] >= 'a' && command2[i] <= 'z'))
                     {
                         sb.Append(command2[i].ToString());
                     }
                     else
                     {
-                        goto Done;  
+                        Console.WriteLine("Argumenti i dyte duhet te jete A-Z,a-z,0-9.");
+                        return;
                     }
                 }
                 command2 = sb.ToString();
                 rsa = new RsaEncryptor();
                         string privateKey = rsa.GetPrivateKey();
                         string publicKey = rsa.GetPublicKey();
-                        File.WriteAllText("keys/"+command2+".xml", privateKey);
-                        File.WriteAllText("keys/"+command2+".pub.xml", publicKey);
                 string privkey = "keys/" + command2 + ".xml";
                 string pubkey = "keys/" + command2 + ".pub.xml";
+                if (File.Exists(privkey)&&File.Exists(pubkey))
+                {
+                    Console.WriteLine("Gabim: Celesi '"+command2+"' ekziston paraprakisht.");
+                    return;
+                }
+                File.WriteAllText("keys/"+command2+".xml", privateKey);
+                File.WriteAllText("keys/"+command2+".pub.xml", publicKey);
+                
                 Console.WriteLine("Eshte krijua celsi privat '"+privkey+"'");
                 Console.WriteLine("Eshte krijua celse publik '"+pubkey+"'");
-                Done:
-                Console.WriteLine("Argumenti i dyte duhet te jete A-Z,a-z,0-9.");
-
+                
             }
-/*            else if (command == "delete-user")
+            else if (command == "delete-user")
             {
 
                 string command2 = args[1];
+                StringBuilder sb = new StringBuilder();
+                for (int i = 0; i < command2.Length; i++)
+                {
+                    if ((command2[i] >= '0' && command2[i] <= '9') || (command2[i] >= 'A' && command2[i] <= 'Z' )|| (command2[i] >= 'a' && command2[i] <= 'z'))
+                    {
+                        sb.Append(command2[i].ToString());
+                    }
+                    else
+                    {
+                        Console.WriteLine("Argumenti i dyte duhet te jete A-Z,a-z,0-9.");
+                        return;
+                    }
+                }
+                command2 = sb.ToString();
                 rsa = new RsaEncryptor();
                 string privateKey = rsa.GetPrivateKey();
                 string publicKey = rsa.GetPublicKey();
                 string privkey = "keys/" + command2 + ".xml";
                 string pubkey = "keys/" + command2 + ".pub.xml";
-
-                string fileKey = privkey;
-                Console.WriteLine(File.Exists(fileKey) ?   : "File does not exist.");
-
-                Console.WriteLine("Eshte larguar celsi privat '" + privkey + "'");
-                Console.WriteLine("Eshte larguar celse publik '" + pubkey + "'");
-
-            }*/
+                if (File.Exists(privkey) && File.Exists(pubkey))
+                {
+                    File.Delete(privkey);
+                    File.Delete(pubkey);
+                    Console.WriteLine("Eshte larguar celsi privat '" + privkey + "'");
+                    Console.WriteLine("Eshte larguar celse publik '" + pubkey + "'");
+                }
+                else
+                {
+                    Console.WriteLine("Gabim: Celesi '"+command2+"' ekziston paraprakisht.");
+                }
+            }
         }
     }
